The BOIN design as a Cumulative-Cohort Design (CCD) subclass

The BOIN design as a Cumulative-Cohort Design (CCD) subclass

The BOIN design as a Cumulative-Cohort Design (CCD) subclass

Details

TODO: Provide references, and note clearly the choice of defaults. Also provide citations to BOIN paper(s).

Super classes

precautionary::Cpe -> precautionary::Ccd -> Boin

Methods

Public methods

Boin$new()

Boin$clone()

Inherited methods

precautionary::Cpe$J()

precautionary::Cpe$bU()

precautionary::Cpe$max_dose()

precautionary::Cpe$path_array()

precautionary::Cpe$path_matrix()

precautionary::Cpe$path_probs()

precautionary::Cpe$path_rx()

precautionary::Cpe$report()

precautionary::Cpe$trace_paths()

precautionary::Ccd$applied()

Method new()

Usage

Boin$new(target, cohort_max, enroll_max)

Arguments

target

Target toxicity rate

cohort_max

Upper bound on dose-wise enrollment

enroll_max

Upper bound on total enrollment

Details

Create a new Boin object.

Returns

A Boin object.

Examples

# TODO

Method clone()

The objects of this class are cloneable with this method.

Usage

Boin$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Boin$new`
## ------------------------------------------------

# TODO