An R6 class encapsulating Cumulative-Cohort Designs

An R6 class encapsulating Cumulative-Cohort Designs

An R6 class encapsulating Cumulative-Cohort Designs

Details

TODO: Explain the hierarchy of model classes, including connections with the executable specifications set forth in exec/prolog/ccd.pl.

References

Ivanova A, Flournoy N, Chung Y. Cumulative cohort design for dose-finding. Journal of Statistical Planning and Inference. 2007;137(7):2316-2327. doi:10.1016/j.jspi.2006.07.009

Liu S, Yuan Y. Bayesian optimal interval designs for phase I clinical trials. J R Stat Soc C. 2015;64(3):507-523. doi:10.1111/rssc.12089

Super class

precautionary::Cpe -> Ccd

Methods

Public methods

Ccd$new()

Ccd$applied()

Ccd$clone()

Inherited methods

precautionary::Cpe$max_dose()

precautionary::Cpe$path_array()

precautionary::Cpe$path_matrix()

precautionary::Cpe$path_probs()

precautionary::Cpe$path_rx()

precautionary::Cpe$report()

precautionary::Cpe$trace_paths()

Method new()

Usage

Ccd$new(escalate, deescalate, eliminate, cohort_max, enroll_max)

Arguments

escalate

Escalation boundary

deescalate

Deescalation boundary

eliminate

Elimination boundary

cohort_max

Upper bound on dose-wise enrollment

enroll_max

Upper bound on total enrollment

Details

Create a new Ccd object.

Returns

A Ccd object.

Examples

# TODO

Method new()

Usage

Ccd$new(escalate, deescalate, eliminate, cohort_max, enroll_max)

Arguments

escalate

Escalation boundary

deescalate

Deescalation boundary

eliminate

Elimination boundary

cohort_max

Upper bound on dose-wise enrollment

enroll_max

Upper bound on total enrollment

Details

Create a new Ccd object.

Returns

A Ccd object.

Examples

# TODO

Method applied()

Usage

Ccd$applied(x, o, last_dose, max_dose, ...)

Arguments

x

A dose-wise vector of toxicity counts

o

A dose-wise vector of non-toxicity counts

last_dose

The most recently given dose, as required to implement cumulative-cohort-based escalation decisions.

max_dose

An upper limit on future dose levels

...

Unused by Ccd; included for superclass method compatibility

Details

Return dose recommendation for given tox/no-tox tallies.

Returns

An object with components:

$stop - logical value indicating whether stop is indicated

$mtd - integer value, the recommended dose

$max_dose - integer value, a dose not to be exceeded henceforth.

Method applied()

Usage

Ccd$applied(x, o, last_dose, max_dose, ...)

Arguments

x

A dose-wise vector of toxicity counts

o

A dose-wise vector of non-toxicity counts

last_dose

The most recently given dose, as required to implement cumulative-cohort-based escalation decisions.

max_dose

An upper limit on future dose levels

...

Unused by Ccd; included for superclass method compatibility

Details

Return dose recommendation for given tox/no-tox tallies.

Returns

An object with components:

$stop - logical value indicating whether stop is indicated

$mtd - integer value, the recommended dose

$max_dose - integer value, a dose not to be exceeded henceforth.

Method clone()

The objects of this class are cloneable with this method.

Usage

Ccd$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

Ccd$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Ccd$new`
## ------------------------------------------------

# TODO