An R6 class encapsulating precalculated CPE for 3+3 design

An R6 class encapsulating precalculated CPE for 3+3 design

An R6 class encapsulating precalculated CPE for 3+3 design

Details

TODO: Explain the hierarchy of model classes, including connections with the executable specifications set forth in exec/prolog/ccd.pl.

Note

This specializes the generic method of Cpe superclass, removing the possibility of updating the state.

References

Korn EL, Midthune D, Chen TT, Rubinstein LV, Christian MC, Simon RM. A comparison of two phase I trial designs. Stat Med. 1994;13(18):1799-1806. doi: 10.1002/sim.4780131802

Super class

precautionary::Cpe -> Cpe3_3

Methods

Public methods

Cpe3_3$new()

Cpe3_3$max_dose()

Cpe3_3$bU()

Cpe3_3$J()

Cpe3_3$trace_paths()

Cpe3_3$path_matrix()

Cpe3_3$path_rx()

Cpe3_3$clone()

Inherited methods

precautionary::Cpe$applied()

precautionary::Cpe$decision_array()

precautionary::Cpe$path_array()

precautionary::Cpe$path_probs()

precautionary::Cpe$report()

Method new()

Usage

Cpe3_3$new(D)

Arguments

D

number of prespecified doses

Details

Create a new Cpe3_3 object.

Returns

A Cpe3_3 object.

Examples

# TODO

Method max_dose()

Usage

Cpe3_3$max_dose()

Details

Query number of doses

Returns

Self (invisibly), unless D is missing, in which case the top dose, an integer, is returned.

Method bU()

Usage

Cpe3_3$bU()

Details

Get the b vector and U matrix

Returns

Named list with components b and U

Method J()

Usage

Cpe3_3$J()

Details

Get the number J of paths

Returns

Integer number of paths

Method trace_paths()

Usage

Cpe3_3$trace_paths(...)

Arguments

...

Ignored

Details

No-op specialization of superclass method

Since Cpe3_3 has cached paths precomputed by Prolog code, it does not need to support this method.

Returns

Self, invisibly

Method path_matrix()

Usage

Cpe3_3$path_matrix()

Details

Refuse superclass method

Returns

An error

Method path_rx()

Usage

Cpe3_3$path_rx()

Details

Refuse superclass method

Returns

An error

Method clone()

The objects of this class are cloneable with this method.

Usage

Cpe3_3$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Cpe3_3$new`
## ------------------------------------------------

# TODO