Execute CRM

Run a CRM trial under given design options for dose-escalation decisions. This is identical to applied_crm, except that it uses precautionary's more performant version of crm.

applied_crm(
  prior,
  scale,
  target,
  tox,
  level,
  no_skip_esc = TRUE,
  no_skip_deesc = TRUE,
  global_coherent_esc = TRUE,
  stop_func = NULL,
  ...
)

Arguments

prior
A vector of prior estimates of toxicity probabilties for the dose levels

scale
The sigma parameter of a lognormal prior (denoted $$g(a)$$ in the original 1990 O'Quigley et al. paper) over the single parameter of the empiric or logistic CRM models.

target
The target DLT rate

tox
A patient-wise vector of 0/1 toxicity indicators

level
A patient-wise vector of dose-level assignments

no_skip_esc
If FALSE, the method will not enforce no skipping of doses in escalation. Default is TRUE.

no_skip_deesc
If FALSE, the method will not enforce no skipping of doses in de-escalation. Default is TRUE.

global_coherent_esc
If FALSE, the method will not enforce global coherent escalation, that is, escalation if the overall rate of toxicity seen at the current dose level is above the target rate. Default is TRUE.

stop_func
An optional argument to provide a function which will utilised alongside the CRM to determine if the trial should be stopped.

...
Additional params addressed to the CRM numerical routines. Presently used to select optional implementations via impl.

Value

An object of class 'mtd', as per dfcrm