Faster Dose Transition Pathways (DTP) calculation

This function reimplements dtpcrm::calculate_dtps with algorithmic improvements that achieve signficant speedups, rendering it suitable for comprehensive enumeration of all paths of a trial.

calculate_dtps(
  next_dose,
  cohort_sizes,
  dose_func = applied_crm,
  ...,
  mc.cores = getOption("mc.cores", 2L)
)

Arguments

next_dose
The root dose of the trial-path tree to be computed

cohort_sizes
An integer vector of cohort sizes, with length equal to the maximum number of cohorts that may be enrolled. (Thus, sum(cohort_sizes) is the trial's maximum enrollment.)

dose_func
An adapter function

...
Ultimately passed to .conduct_dose_finding_cohorts, whence it ends up informing the dose_func. Thanks to the method chaining of R6 class Crm, the role of this argument has been sharply curtailed compared with the original dtpcrm::calculate_dtps. Here, we use it only for selecting optional CRM numerical implementations via impl.

mc.cores
Number of logical cores available for parallelizing DTP computation. Setting this to 1 prevents chunking of the computation.

Value

A data.table listing trial pathways

Details

In accord with the chosen DTP representation of dtpcrm, we pursue a tabular listing of all paths, inclusive of the degeneracies engendered by early stopping. During the DTP computation, however, these degeneracies are recognized and handled efficiently here, avoiding costly reduplication of CRM model runs. The algorithm employs a top-to-bottom scan of the path table, in the lexicographic order of increasing toxicity counts which naturally arises from expand.grid used here and in the original dtpcrm::calculate_dtps. The early termination of any path in the course of this scan may then be carried forward through all subsequent paths which share the same dose assignments and toxicity counts. (In the case of early termination specifically for excessive toxicity, this principle could also be applied 'one level up', possibly capturing some additional efficiencies. But implementing this safely would require greater control over early-termination semantics than dtpcrm --- or indeed perhaps any R package --- seems capable of achieving. So it is not implemented here.)

The substantial efficiencies (3.5x speedup of VIOLA trial DTP) obtained in the above-described scan can readily be preserved under coarse-grained parallelization, on account of the DTP table's recursive structure. If the first cohort has size $$n$$, then it gives rise to $$n+1$$ independent chunks within which the scan logic applies, one chunk for each possible toxicity count in $$\{0,1,...,n\}$$. This principle applies also at the level of the next cohort, and so on. With each such refinement to the chunking of the DTP calculation, some re-work is introduced where the same early termination even is discovered independently in several chunks. But given that few reasonable trial designs will have a substantial number of paths terminating before 6 patients, this re-work should be irrelevant to parallelization of DTP computation over the 4--16 threads of a single modern, multicore processor. (Consider that chunking just the first 2 cohorts of a trial with cohort size 3 yields $(3+1)^2=16$ chunks.)

Author

Adapted by David C. Norris from original dtpcrm::calculate_dtps

Examples

## VIOLA trial
prior.DLT <- c(0.03, 0.07, 0.12, 0.20, 0.30, 0.40, 0.52)
prior.var <- 0.75

stop_func <- function(x) {
  y <- stop_for_excess_toxicity_empiric(x,
                                        tox_lim = 0.3,
                                        prob_cert = 0.72,
                                        dose = 1)
  if(y$stop){
    x <- y
  } else {
    x <- stop_for_consensus_reached(x, req_at_mtd = 12)
  }
}

crm <- Crm$new(skeleton = prior.DLT,
               scale = sqrt(prior.var),
               target = 0.2)$
  stop_func(stop_func)$
  no_skip_esc(TRUE)$
  no_skip_deesc(FALSE)$
  global_coherent_esc(TRUE)

if (interactive()) { # don't overtax CRAN servers
restore <- options(mc.cores = 1) # you can reasonably set as high as detectCores()
dtps <- calculate_dtps(next_dose = 3,
                       cohort_sizes = rep(3, 7),
                       dose_func = crm$applied,
                       impl = "rusti")
options(mc.cores = restore)
}