Override cohorts_of_n to include latent toxicity tolerances

The original function in package escalation recognizes that individual trial participants arrive at distinct times. Building upon this acknowledgment of individuality, this override adds an extra line of code to draw as well a latent toxicity tolerance u_i for each individual participant.

cohorts_of_n(n = 3, mean_time_delta = 1)

Arguments

n

integer, sample arrival times for this many patients.

mean_time_delta

the average gap between patient arrival times. I.e. the reciprocal of the rate parameter in an Exponential distribution.

Value

data.frame with columns u_i and time_deltacontaining respectively the uniformly-distributed latent toxicity tolerance and arrival-time increment for each trial participant.

See also

Examples

cohorts_of_n()
#>          u_i time_delta
#> 1 0.72102296        2.0
#> 2 0.47590148        1.1
#> 3 0.02207596        0.5
cohorts_of_n(n = 10, mean_time_delta = 5)
#>          u_i time_delta
#> 1  0.5847561       15.5
#> 2  0.3608974        9.9
#> 3  0.9086356        7.0
#> 4  0.3651750        6.8
#> 5  0.6213080        1.2
#> 6  0.7596542        2.2
#> 7  0.3983997        6.3
#> 8  0.7596544       20.5
#> 9  0.8400901       14.1
#> 10 0.3869112        4.8