A package-local (as-yet, unexported) test harness adapted from dfcrm::crm().

for various performance tuning experiments. The impl arg allows selection of various alternative implementations:

'rusti' substitutes integrands crmh, crmht, crmht2 written in Rust

'dfcrm' is the original as implemented in package dfcrm.

crm(
  prior,
  target,
  tox,
  level,
  n = length(level),
  dosename = NULL,
  include = 1:n,
  pid = 1:n,
  conf.level = 0.9,
  method = "bayes",
  model = "empiric",
  intcpt = 3,
  scale = sqrt(1.34),
  model.detail = TRUE,
  patient.detail = TRUE,
  var.est = TRUE,
  impl = c("rusti", "ruste", "dfcrm")
)

Arguments

prior

The CRM skeleton: dose-wise prior probabilities of toxicity

target

Target toxicity rate

tox

A patient-wise vector of toxicity counts

level

A patient-wise vector of dose level assignments

n

The number of patients enrolled

dosename

Optional designators for the doses

include

Index of patients to include

pid

Vector of patient ID labels

conf.level

Used to assign upper and lower bounds on predicted probability of toxicity, which in turn may be referenced in escalation, deescalation and stopping decisions.

method

Estimation method:

model

Presently, only the ‘empiric’ (or ‘power’) model has a Rust likelihood implementation.

intcpt

Intercept for ‘logistic’ model

scale

Sigma parameter of prior on beta parameter

model.detail

If FALSE, the model content of an mtd object will not be displayed. Default is TRUE.

patient.detail

If FALSE, patient summary of an mtd object will not be displayed. Default is TRUE.

var.est

If TRUE, variance of the estimate of the model parameter and probability/confidence interval for the dose-toxicity curve will be computed

impl

Switch between 'rusti' and 'dfcrm' implementations. Currently the 'rusti' option is implemented only for the Bayes method of the empirical (‘power’) model. An experimental 'ruste' implementation is in the works.

Author

Adapted by David C. Norris, from Ken Cheung's dfcrm