Extend an existing simulation, using one of several stopping criteria

A trial simulation carried through a predetermined number of replications may not achieve desired precision as judged by Monte Carlo standard errors (MCSE) of estimated toxicity counts. This method enables simulations of class c('precautionary','simulations') to be extended until a given level of precision has been achieved on expected counts of enrollment and DLTs, or (optionally) for a fixed additional number of simulations.

extend(sims, num_sims = NULL, target_mcse = 0.05)

Arguments

sims
An existing object of class c('precautionary','simulations')

num_sims
Optionally, a fixed number of additional replications to accumulate

target_mcse
Optionally, an MCSE constraint to be imposed on expected counts of DLTs, non-DLTs, and Total enrollment.

Value

An extended simulation of same class as sims.

Note

The MCSE constraint is imposed during trial simulation, at which point only binary toxicities are available. Thus, as a practical matter, extend can target MCSEs only for DLTs, non-DLTs and Total enrollment. The subsequent subdivision of these categories during trial summary (at which point the ordinalizer comes into play along with its parameters) thus may generate expected counts with MCSEs exceeding target_mcse. In practice, however, this tends to affect the estimated counts only for the lowest toxicity grades---those of least concern from a trial-safety perspective.

See also