Override escalation:::phase1_sim to incorporate latent toxicity tolerances

Override escalation:::phase1_sim to incorporate latent toxicity tolerances

phase1_sim(
  selector_factory,
  true_prob_tox,
  sample_patient_arrivals = function(df) cohorts_of_n(n = 3, mean_time_delta = 1),
  previous_outcomes = "",
  next_dose = NULL,
  i_like_big_trials = FALSE,
  return_all_fits = FALSE
)

Arguments

selector_factory
A selector_factory object

true_prob_tox
A vector of toxicity probabilities for the doses defined in selector_factory

sample_patient_arrivals
A function implementing an arrivals process for trial enrollment

previous_outcomes
This may or may not apply in applications of package precautionary

next_dose
Undocumented

i_like_big_trials
I didn't choose this parameter name

return_all_fits
Don't do this