Plan

Plan

Version 0.2.6-1

Improve feedback during long calculations

Set J = '...' pending first progress report

Analyze where Cpe$path_array() spends time

Give suitable feedback on path_array progress

Version 0.2.7

Demote escalation to a 'Suggests'

Thoroughly rewrite 'Intro' and 'FDA Clinical Hold' vignettes

Version 0.2.8

BOIN recs via isotonic regression

Testthat CPE matches BOIN::get.oc() sim

Implement CRM logistic model

Version 0.2.9

Implement TITE CRM?

Is TITE even amenable to path enumeration?

If not, there may be little sense in implementing it here!

If TITE is truly off the table, can existing numerics be sped up further?

OTOH, implementing TITE would help place the numerics for enumerable designs in context.

Version 0.3.0

Stop exposing the impl parameter

Version 0.3.1

Native Rust CPE

Dependencies

Eliminate dependence on/adherence to dtpcrm layout

Package dtpcrm makes many design decisions unsuited to comprehensive enumeration of whole trials.

Allowing the Crm class to evolve along now-'natural' lines probably makes a suitable 'plan' for now.

Expunging superseded code/dependencies ASAP will facilitate this evolution.

Even the term 'DTP' seems not quite right anymore for a complete path enumeration (CPE?), and should be abandoned.

Eliminate dependence on 'escalation' and 'dfcrm'

With the special emphasis on speed created by CPE, I now must implement all underlying trial designs in Rust.

What may be retained in each case are 'Suggests:'-type relations, wherever regression tests are helpful -- esp. wrt dfcrm.

Fast CRM

Benchmark; try mul_add()'s

Document

Add examples to the documented Rust functions

Add a vignette applying DTP to mTPI and BOIN

Expose some visual numerics checks, via vignette or documented function

Refactor

Remove the $safety component of exact trials?

Perhaps this ought to be calculated 'on the fly' by the summary method.

On-the-fly calculation would postpone use of ordinalizer, in keeping with the pattern established already for (non-exact) simulations.

Should summary(EXACT)$safety bear class 'safetytab'?

Example for 'as.data.table.exact'

What is role of G function in exact.R?

Extend

Implement exact 3+3 variant with allow_deescalation=FALSE

Implement rolling 6

Allow an accelerated titration phase

Note that this requires access to graded toxicities at simulation time, and therefore constitutes a substantial challenge to the generalizability of this software design.

Index sims$fits to exact outcomes in A[[D]] where appropriate

See the haystack function in exact.R

Robustify

Tests comparing results from multiple CRAN packages