Get a function that simulates dose-escalation trials using latent u_i

Overrides simulation_function.tox_selector_factory to return a phase1_sim() that employs latent u_i in place of the version native to package escalation, which merely invokes rbinom.

# S3 method for u_i
simulation_function(selector_factory)

Arguments

selector_factory
Presently, this must be a tox_selector_factory; no equivalent for simulation_function.derived_dose_selector_factory is yet implemented.

Details

This function is exported for the purpose of effecting this override, and is not meant to be invoked directly by the user.