An R6 class encapsulating CRM models This class wraps the functionality of package dfcrm, adding efficient Rust implementations of some numerical routines.

An R6 class encapsulating CRM models

This class wraps the functionality of package dfcrm, adding efficient Rust implementations of some numerical routines.

An R6 class encapsulating CRM models

This class wraps the functionality of package dfcrm, adding efficient Rust implementations of some numerical routines.

Details

Syntactically, the method chaining supported by R6 classes makes the invocation of CRM models more transparent. The mutability conferred by reference semantics enables memoization (caching) of results, which can speed up DTP calculations significantly.

Presently, this class supports only the 'empiric' (aka 'power') model. But it is hoped that inheritance will assist in rendering other models implemented in package dfcrm clearly, with code reuse.

References

Norris DC. What Were They Thinking? Pharmacologic priors implicit in a choice of 3+3 dose-escalation design. arXiv:2012.05301 [stat.ME]. December 2020. https://arxiv.org/abs/2012.05301

See also

Public fields

performance

A vector used for vetting performance

Methods

Public methods

Crm$new()

Crm$skeleton()

Crm$dontcache()

Crm$report()

Crm$stop_func()

Crm$no_skip_esc()

Crm$no_skip_deesc()

Crm$global_coherent_esc()

Crm$conf_level()

Crm$observe()

Crm$tally()

Crm$est()

Crm$applied()

Crm$trace_paths()

Crm$path_matrix()

Crm$path_array()

Crm$path_probs()

Crm$path_rx()

Crm$clone()

Method new()

Usage

Crm$new(skeleton, scale = sqrt(1.34), target)

Arguments

skeleton

CRM skeleton

scale

Sigma parameter of prior on beta parameter

target

Target toxicity rate

Details

Create a new Crm object.

Returns

A Crm object.

Examples

# An example verbatim from dfcrm::crm()
prior <- c(0.05, 0.10, 0.20, 0.35, 0.50, 0.70)
target <- 0.2
level <- c(3, 4, 4, 3, 3, 4, 3, 2, 2, 2)
y     <- c(0, 0, 1, 0, 0, 1, 1, 0, 0, 0)
s <- sqrt(1.34)
old <- dfcrm::crm(prior, target, y, level)
new <- Crm$new(skeleton = prior, target = target)$
         dontcache()$
         observe(level, y)$
         est(impl="rusti", abbrev=FALSE)

Method skeleton()

Usage

Crm$skeleton(skeleton)

Arguments

skeleton

A numeric vector to set as the model skeleton.

Details

Set or query CRM skeleton

Returns

Self (invisibly), unless skeleton is missing, in which case the skeleton, a numeric vector, is returned.

Method dontcache()

Usage

Crm$dontcache()

Details

Set private cache to NULL; useful for performance testing

Returns

Self, invisibly

Method report()

Usage

Crm$report()

Details

Report lifetime duty & performance statistics

Returns

A named vector summarizing lifetime duty and performance

Method stop_func()

Usage

Crm$stop_func(sfunc)

Arguments

sfunc

A function taking mtd objects to mtd objects, attaching suitable stopping information

Details

Set the stopping function

Returns

Self, invisibly

Method no_skip_esc()

Usage

Crm$no_skip_esc(tf)

Arguments

tf

An atomic logical value, TRUE or FALSE

Details

Set the no_skip_esc behavior

Returns

Self, invisibly

Method no_skip_deesc()

Usage

Crm$no_skip_deesc(tf)

Arguments

tf

An atomic logical value, TRUE or FALSE

Details

Set the no_skip_deesc behavior

Returns

Self, invisibly

Method global_coherent_esc()

Usage

Crm$global_coherent_esc(tf)

Arguments

tf

An atomic logical value, TRUE or FALSE

Details

Set the global_coherent_esc behavior

Returns

Self, invisibly

Method conf_level()

Usage

Crm$conf_level(conf)

Arguments

conf

A numeric confidence less than 1.0

Details

Set the required confidence level for escalation decisions

Returns

Self, invisibly

Method observe()

Usage

Crm$observe(level, tox)

Arguments

level

A patient-wise vector of dose assignments

tox

A patient-wise vector of 0/1 toxicity assessments

Details

Set patient-wise toxicity observations

Returns

Self, invisibly

Method tally()

Usage

Crm$tally(x, o)

Arguments

x

A dose-wise vector of toxicity counts

o

A dose-wise vector of non-toxicity counts

Details

Set dose-wise toxicity observations

Returns

Self, invisibly

Method est()

Usage

Crm$est(impl, abbrev = TRUE)

Arguments

impl

A string choosing the low-level implementation to use. Possible values include "dfcrm", "rusti" and "ruste".

abbrev

Logical; if TRUE (the default), an abbreviated mtd object is returned to save execution time. If FALSE, a complete object is returned, suitable for regression testing against package dfcrm.

Details

Set patient-wise toxicity observations

Returns

An object of class mtd as per package dfcrm

Method applied()

Usage

Crm$applied(x, o, last_dose = NA, ...)

Arguments

x

A dose-wise vector of toxicity counts

o

A dose-wise vector of non-toxicity counts

last_dose

The most recently given dose, as required to implement the global_coherent_esc=TRUE behavior

...

Parameters passed to Crm$esc(), enabling passthru of required impl parameter and optional abbrev flag.

Details

Return dose recommendation for given tox/no-tox tallies.

This function caches results, which greatly saves computation time in DTP calculations. (It yields approximately a 5x speedup for the VIOLA trial example!)

Returns

An object of class mtd as per package dfcrm, or possibly an abbreviated version of such object as returned by method Crm$est().

Method trace_paths()

Usage

Crm$trace_paths(root_dose, cohort_sizes, ..., unroll = 4)

Arguments

root_dose

The starting dose for tree of paths

cohort_sizes

Integer vector giving sizes of future cohorts, its length being the maximum number of cohorts to look ahead.

...

Parameters passed ultimately to Crm$esc(), enabling passthru of its required impl parameter.

unroll

Integer; how deep to unroll path tree for parallelism

Details

Compute trial paths forward from current tally

The computed paths are saved in a private field, from which variously formatted results may be obtained by accessor functions.

Returns

Self, invisibly

Method path_matrix()

Usage

Crm$path_matrix()

Details

Return computed trial paths in matrix form

Returns

An integer matrix with the same column layout as the DTP tables of dtpcrm. That is, there is a D0 column followed by paired Tc, Dc columns giving the toxicity count for cohort c and the resulting dose recommendation yielded by cohort c -- which is generally the recommendation for cohort c+1.

Method path_array()

Usage

Crm$path_array(condense = TRUE)

Arguments

condense

Logical value; if FALSE, the returned array has its cohorts indexed trial-wise instead of dose-wise. This inflates the array more than needed for the matrix computations it must support (observe that in Norris2020c Eq. (4), the c index is eliminated already by summation), but enables the sequence of events along a path to be read off directly if this is required e.g. for visualization or debugging. Default is TRUE.

Details

Return computed trial paths in a 3D array

Returns

For the j'th path, the C*D matrix T[j,,] gives the number of toxicities T[j,c,d] occurring in the c'th cohort for dose d. In case condense=FALSE, see above.

Method path_probs()

Usage

Crm$path_probs(probs.DLT)

Arguments

probs.DLT

Numeric vector of DLT probabilities for the design's pre-specified doses.

Details

Path probabilities for given dose-wise DLT probs

The design's paths must already have been completely enumerated by trace_paths, and the associated array T, matrix U and vector b computed by a call to path_array.

Returns

A vector of probabilities for the enumerated paths

Method path_rx()

Usage

Crm$path_rx()

Details

Vector of path-wise final dose recommendations

The design's paths must already have been completely enumerated by trace_paths. This method is a temporizing measure, bridging to a new format for the return value of path_matrix, possibly a data.table where the dose recs would be a column.

Returns

Integer vector of final dose recs on the enumerated paths

Method clone()

The objects of this class are cloneable with this method.

Usage

Crm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `Crm$new`
## ------------------------------------------------

# An example verbatim from dfcrm::crm()
prior <- c(0.05, 0.10, 0.20, 0.35, 0.50, 0.70)
target <- 0.2
level <- c(3, 4, 4, 3, 3, 4, 3, 2, 2, 2)
y     <- c(0, 0, 1, 0, 0, 1, 1, 0, 0, 0)
s <- sqrt(1.34)
old <- dfcrm::crm(prior, target, y, level)
new <- Crm$new(skeleton = prior, target = target)$
         dontcache()$
         observe(level, y)$
         est(impl="rusti", abbrev=FALSE)