Simulate trials defined via package escalation

An S4 generic method providing a more abstract interface to trial simulation than the simulate_trials function of package escalation. This abstraction is needed to support simulations in which escalation's simple vectors of ‘true toxicity probabilities’ are replaced by precautionary's more realistic toxicity-distribution generators.

simulate_trials(
  selector_factory,
  num_sims,
  true_prob_tox,
  true_prob_eff = NULL,
  ...
)

# S4 method for selector_factory,numeric,hyper_mtdi_distribution,ANY
simulate_trials(
  selector_factory,
  num_sims,
  true_prob_tox,
  true_prob_eff = NULL,
  ...
)

# S4 method for selector_factory,numeric,mtdi_distribution,ANY
simulate_trials(
  selector_factory,
  num_sims,
  true_prob_tox,
  true_prob_eff = NULL,
  ...
)

# S4 method for exact,missing,mtdi_distribution,ANY
simulate_trials(
  selector_factory,
  num_sims,
  true_prob_tox,
  true_prob_eff = NULL,
  ...
)

# S4 method for exact,numeric,mtdi_distribution,missing
simulate_trials(
  selector_factory,
  num_sims,
  true_prob_tox,
  true_prob_eff = NULL,
  ...
)

# S4 method for exact,numeric,hyper_mtdi_distribution,missing
simulate_trials(
  selector_factory,
  num_sims,
  true_prob_tox,
  true_prob_eff = NULL,
  ...
)

Arguments

selector_factory
An object of S3 class selector_factory

num_sims
Number of simulations to run

true_prob_tox
A generator of toxicity distributions

true_prob_eff
Provided for compatibility with simulate_trials

...
Passed to subroutines

Details

If invoked interactively with num_sims > 10, then a txtProgressBar is displayed in the console. The condition on num_sims has the useful side effect of allowing this function to be invoked iteratively by extend (with num_sims = 10) without the nuisance of nested progress bars.

Examples

old <- options(dose_levels = c(0.5, 1, 2, 4, 6, 8))
mtdi_gen <- hyper_mtdi_lognormal(CV = 1
                                 , median_mtd = 6, median_sdlog = 0.5
                                 , units="mg/kg")
num_sims <- ifelse(interactive()
, 300
, 15 # avoid taxing CRAN servers
)
hsims <- get_three_plus_three(num_doses = 6,
                              allow_deescalate = TRUE) %>%
  simulate_trials(
    num_sims = num_sims
  , true_prob_tox = mtdi_gen)
summary(hsims, ordinalizer=NULL) # vanilla summary with binary toxicity
#> # A tibble: 7 x 7
#>   dose   `dose (mg/kg)`   tox     n true_prob_tox prob_recommend prob_administer
#>   <ord>           <dbl> <dbl> <dbl>         <dbl>          <dbl>           <dbl>
#> 1 NoDose            0   0       0               0         0               0     
#> 2 1                 0.5 0       3.2             0         0.0667          0.178 
#> 3 2                 1   0.4     4.2             0         0.2             0.233 
#> 4 3                 2   0.8     4.2             0         0.333           0.233 
#> 5 4                 4   0.8     3.2             0         0.133           0.178 
#> 6 5                 6   0.467   2               0         0.2             0.111 
#> 7 6                 8   0.4     1.2             0         0.0667          0.0667
summary(hsims, ordinalizer = function(dose, r0 = sqrt(2))
  c(Gr1=dose/r0^2, Gr2=dose/r0, Gr3=dose, Gr4=dose*r0, Gr5=dose*r0^2)
)
#> $escalation
#> # A tibble: 7 x 7
#>   dose   `dose (mg/kg)`   tox     n true_prob_tox prob_recommend prob_administer
#>   <ord>           <dbl> <dbl> <dbl>         <dbl>          <dbl>           <dbl>
#> 1 NoDose            0   0       0               0         0               0     
#> 2 1                 0.5 0       3.2             0         0.0667          0.178 
#> 3 2                 1   0.4     4.2             0         0.2             0.233 
#> 4 3                 2   0.8     4.2             0         0.333           0.233 
#> 5 4                 4   0.8     3.2             0         0.133           0.178 
#> 6 5                 6   0.467   2               0         0.2             0.111 
#> 7 6                 8   0.4     1.2             0         0.0667          0.0667
#> 
#> $safety
#>   None  Gr1  Gr2  Gr3  Gr4  Gr5  Total
#>     11    2    2    1    1    1     18
#> 
#> $toxTab
#>     Tox
#> rep  None Gr1 Gr2 Gr3 Gr4 Gr5 Total
#>   1    16   3   3   2   0   0    24
#>   2    12   2   0   1   1   2    18
#>   3     8   2   0   1   1   0    12
#>   4    12   4   2   2   1   0    21
#>   5     9   1   2   1   1   1    15
#>   6     4   3   2   2   1   0    12
#>   7     9   1   2   0   2   1    15
#>   8    12   2   1   3   0   0    18
#>   9     8   2   1   0   3   1    15
#>   10   13   3   2   0   0   0    18
#>   11   11   4   7   2   0   0    24
#>   12   14   3   2   1   1   3    24
#>   13    9   4   2   3   0   0    18
#>   14   12   0   0   0   2   1    15
#>   15   15   1   2   2   1   0    21
#> 
hsims <- hsims %>% extend(num_sims = num_sims)
summary(hsims, ordinalizer = function(dose, r0 = sqrt(2))
  c(Gr1=dose/r0^2, Gr2=dose/r0, Gr3=dose, Gr4=dose*r0, Gr5=dose*r0^2)
)$safety
#>   None  Gr1  Gr2  Gr3  Gr4  Gr5  Total
#>     11    2    2    1    1    1     18
# Set a CRM skeleton from the average probs in above simulation
num_sims <- ifelse(interactive()
, 16
,  4  # avoid taxing CRAN servers
)
get_dfcrm(skeleton = hsims$avg_prob_tox
         ,target = 0.25
         ) %>% stop_at_n(n = 24) %>%
  simulate_trials(
    num_sims = num_sims
  , true_prob_tox = mtdi_gen
  ) -> crm_hsims
summary(crm_hsims
, ordinalizer = function(MTDi, r0 = sqrt(2))
    MTDi * r0^c(Gr1=-2, Gr2=-1, Gr3=0, Gr4=1, Gr5=2)
)
#> $escalation
#> # A tibble: 7 x 7
#>   dose   `dose (mg/kg)`   tox     n true_prob_tox prob_recommend prob_administer
#>   <ord>           <dbl> <dbl> <dbl>         <dbl>          <dbl>           <dbl>
#> 1 NoDose            0    0     0                0           0             0     
#> 2 1                 0.5  0     3                0           0             0.125 
#> 3 2                 1    0     0                0           0             0     
#> 4 3                 2    0.25  2.25             0           0             0.0938
#> 5 4                 4    0.5   3.75             0           0.5           0.156 
#> 6 5                 6    2.25  8.25             0           0.25          0.344 
#> 7 6                 8    2.25  6.75             0           0.25          0.281 
#> 
#> $safety
#>   None  Gr1  Gr2  Gr3  Gr4  Gr5  Total
#>      9    4    6    3    2    1   24.0
#> 
#> $toxTab
#>    Tox
#> rep None Gr1 Gr2 Gr3 Gr4 Gr5 Total
#>   1    3   4  10   6   1   0    24
#>   2    9   3   9   2   1   0    24
#>   3   12   3   4   1   2   2    24
#>   4   11   5   2   2   2   2    24
#> 
options(old)
old <- options(dose_levels = c(2, 6, 20, 60, 180, 400))
mtdi_dist <- mtdi_lognormal(CV = 0.5
                           ,median = 140
                           ,units = "ng/kg/week")
num_sims <- ifelse(interactive()
, 100
,  10  # avoid taxing CRAN servers
)
sims <- get_three_plus_three(num_doses = 6) %>%
  simulate_trials(
    num_sims = num_sims
  , true_prob_tox = mtdi_dist)
# Now set a proper ordinalizer via options():
options(ordinalizer = function(dose, r0) {
  c(Gr1=dose/r0^2, Gr2=dose/r0, Gr3=dose, Gr4=dose*r0, Gr5=dose*r0^2)
})
summary(sims, r0=2)
#> $escalation
#> # A tibble: 7 x 7
#>   dose   `dose (ng/kg/week)`   tox     n true_prob_tox prob_recommend
#>   <ord>                <dbl> <dbl> <dbl>         <dbl>          <dbl>
#> 1 NoDose                   0   0     0        0                   0  
#> 2 1                        2   0     3        1.19e-19            0  
#> 3 2                        6   0     3        1.30e-11            0  
#> 4 3                       20   0     3        1.90e- 5            0.1
#> 5 4                       60   0.4   3.9      3.64e- 2            0.9
#> 6 5                      180   2.7   3.3      7.03e- 1            0  
#> 7 6                      400   0     0        9.87e- 1            0  
#> # … with 1 more variable: prob_administer <dbl>
#> 
#> $safety
#>   None  Gr1  Gr2  Gr3  Gr4  Gr5  Total
#>      9    2    2    3    0  0.0     16
#> 
#> $toxTab
#>     Tox
#> rep  None Gr1 Gr2 Gr3 Gr4 Gr5 Total
#>   1     9   2   3   3   1   0    18
#>   2     9   1   2   3   0   0    15
#>   3    10   2   1   1   1   0    15
#>   4     9   3   0   2   1   0    15
#>   5     9   1   2   2   1   0    15
#>   6     9   2   2   2   0   0    15
#>   7     8   2   3   2   0   0    15
#>   8     9   2   3   4   0   0    18
#>   9     9   4   1   3   1   0    18
#>   10    9   2   3   4   0   0    18
#> 
# Set a CRM skeleton from the average probs in above simulation
get_dfcrm(skeleton = sims$true_prob_tox
         ,target = 0.25
         ) %>% stop_at_n(n = 24) %>%
  simulate_trials(
    num_sims = 20
  , true_prob_tox = mtdi_dist
  ) -> crm_sims
summary(crm_sims
, ordinalizer = function(MTDi, r0 = sqrt(2))
    MTDi * r0^c(Gr1=-2, Gr2=-1, Gr3=0, Gr4=1, Gr5=2)
)
#> $escalation
#> # A tibble: 7 x 7
#>   dose   `dose (ng/kg/week)`   tox     n true_prob_tox prob_recommend
#>   <ord>                <dbl> <dbl> <dbl>         <dbl>          <dbl>
#> 1 NoDose                   0  0     0         0                     0
#> 2 1                        2  0     3         1.19e-19              0
#> 3 2                        6  0     0         1.30e-11              0
#> 4 3                       20  0     0         1.90e- 5              0
#> 5 4                       60  0.85 18.8       3.64e- 2              1
#> 6 5                      180  1.7   2.25      7.03e- 1              0
#> 7 6                      400  0     0         9.87e- 1              0
#> # … with 1 more variable: prob_administer <dbl>
#> 
#> $safety
#>   None  Gr1  Gr2  Gr3  Gr4  Gr5  Total
#>     15    4    2    1    1    0   24.0
#> 
#> $toxTab
#>     Tox
#> rep  None Gr1 Gr2 Gr3 Gr4 Gr5 Total
#>   1    13   5   3   1   2   0    24
#>   2    13   8   1   2   0   0    24
#>   3    17   4   2   1   0   0    24
#>   4    14   7   2   1   0   0    24
#>   5    14   5   2   3   0   0    24
#>   6    16   3   2   2   1   0    24
#>   7    12   4   5   0   3   0    24
#>   8    18   2   1   0   3   0    24
#>   9    15   2   2   2   2   1    24
#>   10   15   1   5   0   3   0    24
#>   11   17   4   2   1   0   0    24
#>   12   17   3   0   3   1   0    24
#>   13   16   4   1   1   2   0    24
#>   14   13   7   1   1   2   0    24
#>   15   12   5   4   2   1   0    24
#>   16   17   4   2   1   0   0    24
#>   17   19   2   2   1   0   0    24
#>   18   17   5   1   1   0   0    24
#>   19   15   3   2   1   1   2    24
#>   20   15   4   2   1   2   0    24
#> 
if (interactive()) { # don't overtax CRAN servers
crm_sims <- crm_sims %>% extend(target_mcse = 0.1)
summary(crm_sims
, ordinalizer = function(MTDi, r0 = sqrt(2))
    MTDi * r0^c(Gr1=-2, Gr2=-1, Gr3=0, Gr4=1, Gr5=2)
)$safety
}
options(old)
old <- options(
  dose_levels = c(0.5, 1, 2, 4, 6),
  ordinalizer = function(MTDi, r0 = 1.5) {
    MTDi * r0 ^ c(Gr1=-2, Gr2=-1, Gr3=0, Gr4=1, Gr5=2)
  })
mtdi_dist <- mtdi_lognormal(CV = 2
                            ,median = 5
                            ,units = "mg/kg")
design <- get_three_plus_three(num_doses = 5, allow_deescalate = TRUE)
# Note use of wrapper function 'exact'; see ?precautionary::exact.
exact(design) %>% simulate_trials(true_prob_tox = mtdi_dist) -> EXACT
stopifnot(all.equal(1, sum(exp(EXACT$log_pi))))
summary(EXACT)$safety
#>       
#>            None      Gr1      Gr2      Gr3       Gr4      Gr5    Total
#>   [1,] 10.04956 1.591247 1.364949 1.088902 0.8066095 1.317687 16.21896
if (interactive()) { # don't overtax CRAN servers
# Compare with result of discrete-event simulation
design %>% simulate_trials(
  num_sims = 200
  , true_prob_tox = mtdi_dist
) -> SIMS
summary(SIMS)$safety
}
options(old)
old <- options(dose_levels = c(0.5, 1, 2, 4, 6, 8))
mtdi_gen <- hyper_mtdi_lognormal(CV = 1
                                 , median_mtd = 6, median_sdlog = 0.5
                                 , units="mg/kg")
options(ordinalizer = function(dose, r0 = sqrt(2))
  c(Gr1=dose/r0^2, Gr2=dose/r0, Gr3=dose, Gr4=dose*r0, Gr5=dose*r0^2)
)
design <- get_three_plus_three(num_doses = 6, allow_deescalate = TRUE)
ehsims <- simulate_trials(
  exact(design)
  , num_sims = 50
  , true_prob_tox = mtdi_gen
)
summary(ehsims)$safety
#>                             None       Gr1        Gr2        Gr3        Gr4
#> Expected participants 11.6858336 2.0724898 1.66385433 1.26822795 0.76452350
#> MCSE                   0.3485948 0.1133285 0.08377771 0.05885133 0.02724011
#>                              Gr5      Total
#> Expected participants 0.92689126 18.3818204
#> MCSE                  0.06611548  0.3381676
options(old)